Replace old "rollup" plugin creation.
Replacing the old "rollup" with a more accurate plugin. The old plugin
would have the findings from the latest plugin, which might be lower
CVSS score or criticality than some of the older plugins, which is not
correct. The older, higher scores are more relevant. So we'll take the
best of each plugin, combine those into this plugin.